complete lab3

+++ inf_sec3/src/main/java/MyRandom.java
@@ -5,22 +5,22 @@ public class MyRandom {
 
     public MyRandom() {}
 
-    public List<Double> generate(int count) {
+    public List<Long> generate(int count) {
         MyParams params = new MyParams();
-        List<Double> result = new ArrayList<>();
-        result.add(params.getU0());
+        List<Long> result = new ArrayList<>();
+        result.add((long) params.getU0());
         for (int i = 1; i < count; i++) {
-            result.add((params.getA() * result.get(i - 1) + params.getC()) % params.getM());
+            result.add((long) ((params.getA() * result.get(i - 1) + params.getC()) % params.getM()));
         }
 
         return result;
     }
 
-    public List<Double> generate(int count, MyParams params) {
-        List<Double> result = new ArrayList<>();
-        result.add(params.getU0());
+    public List<Long> generate(int count, MyParams params) {
+        List<Long> result = new ArrayList<>();
+        result.add((long) params.getU0());
         for (int i = 1; i < count; i++) {
-            result.add((params.getA() * result.get(i - 1) + params.getC()) % params.getM());
+            result.add((long) ((params.getA() * result.get(i - 1) + params.getC()) % params.getM()));
         }
 
         return result;
